# server/swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is a Node-Express app called "WeConnect", it provides a platform that brings businesses and individuals together.    
    There is no need for Authorization for these endpoints yet...
  version: 1.0.0
  title: WEconnect App
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: morayodeji@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: ayo-weconnect-dummy.herokuapp.com
basePath: /api/v1
...
schemes:
  - https
  - http
paths:
...
  /businesses:
    post:
      tags:
        - business
      summary: Add a new business to the app
      description: 'creates a business for a user'
      operationId: createBusiness
      consumes:
        - application/json
      security:
        - APIKeyHeader: []
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Business that needs to be added to the app
          schema:
            $ref: '#/definitions/Business'
      responses:
        '400':
          description: Empty fields found.
        '201':
          description: created successfully
    get:
      tags:
        - business
      summary: Returns all business or according to query parameters (Location and category)
      description: '"```/businesses?location=<location>" to query by location and "```/businesses?category=<category>" to query by category' 
      operationId: getAllBusiness
      produces: []
      responses:
        '200':
          description: successful operation
          schema:
            type: object
  '/businesses/{businessId}':
    put:
      tags:
        - business
      summary: Update an existing business
      description: ''
      operationId: modifyBusiness
      consumes:
        - application/json
      security:
        - APIKeyHeader: []
      produces:
        - application/json
      parameters:
        - name: businessId
          in: path
          description: Business id to edit
          required: true
          type: integer
          format: int64
      responses:
        '404':
          description: Business not found
        '200':
          description: Edit successful
    delete:
      tags:
        - business
      summary: Deletes a business
      description: ''
      operationId: deleteBusiness
      produces:
        - application/json
      security:
        - APIKeyHeader: []
      parameters:
        - name: businessId
          in: path
          description: Business id to delete
          required: true
          type: integer
          format: int64
      responses:
        '404':
          description: Business not found
        '200':
          description: delete successful
    get:
      tags:
        - business
      summary: Finds a business by Id
      description: ''
      operationId: getoneBusiness
      produces:
        - application/json
      parameters:
        - name: businessId
          in: path
          description: Business that will be gotten
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
        '404':
          description: Business not found
...
securityDefinitions:
   # X-API-Key: abcdef12345
   APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key
definitions:
...
  Business:
    type: object
    required:
      - name
      - phoneno
      - location
      - category
      - services
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: Abshat Tech
      address:
        type: string
      website:
        type: string
      phoneno:
        type: string
      details:
        type: string
      location:
        type: string
      category:
        type: string
      services:
        type: string
...